pyproject.toml
[tool.poetry]
name = "lightdetect"
version = "0.1.0"
description = ""
authors = ["mward25 <milesward2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.9.0"
zigpy-cli = "1.0.4"
#zigpy = "0.63.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
